Several fixes for Bestow
