+++ src/mahoji/lib/abstracted_commands/lmsCommand.ts
@@ -35,7 +35,7 @@ export async function lmsCommand(
 		if (!itemToBuy) {
 			return "That's not a valid item you can buy.";
 		}
-		const quantity = options.buy.quantity ?? 1;
+		const quantity = Math.max(1, options.buy.quantity ?? 1);
 		const cost = itemToBuy.cost ? itemToBuy.cost * quantity : itemToBuy.cost;
 		if (cost && stats.points < cost) {
 			return `You don't have enough points. ${quantity}x ${itemToBuy.item.name} costs ${cost}, but you have ${stats.points}.`;
