+++ src/mahoji/lib/abstracted_commands/mahoganyHomesCommand.ts
@@ -114,6 +114,7 @@ export async function mahoganyHomesBuyCommand(user: MUser, input = '', quantity?
 	if (!quantity) {
 		quantity = 1;
 	}
+	quantity = Math.max(1, quantity);
 
 	const { item, cost } = buyable;
 	const balance = user.user.carpenter_points;
